google_cloud_app/cloud_run_app/requirements.txt
flask
google-cloud-bigquery
pandas
db-dtypes